Keep unique master bbdb version number in configure.ac
Other necessary copies of the version number are updated at
configuration time.  This is done by the file lisp/bbdb-version.el
generated from lisp/bbdb-version.el.in, which is then loaded in
bbdb.el.  Due to this, a variety of dependencies related to
bbdb-version.el needed to be added, and some code in configure.ac and
the support m4 macro that attempted to extract the version number from
lisp/bbdb.el could be removed.

NOTE: There is, to my knowledge, *no legal way* to extract the version
number from somewhere else and smoosh it into configure.ac.  The only
way to allow the version number to be changed manually by changing it
in one single place is to have that place be the AC_INIT line in
configure.ac.  Trying to put it in an m4 macro will fail in obscure
ways, like giving distribution tarballs the wrong name.

But if you *did* want to extract it from the lisp sources, knowing
full well that this will not only not work, but will also anger the
gods of autotools who will rain woe upon your most precious
configuration scripts for seven generations of automake, a reasonable
fragment of shell code would be:

${EMACS_PROG} -batch --no-site-file --no-init-file \
       --load lisp/bbdb.el \
       -eval '(princ (if (string-match "^[$]Date[:] \\([0-9/]+\\) [0-9:]+ [$]$" bbdb-version-date) (concat bbdb-version "." (subst-char-in-string ?/ ?. (match-string 1 bbdb-version-date))) bbdb-version))'